Refactored, extracted math functions to separated module
