+++ code/kainov-oleg/src/test_fraction.py
@@ -1,6 +1,6 @@
 import unittest
 
-from fraction import Fraction, GCD, LCM, InvalidFractionError
+from fraction import Fraction, InvalidFractionError
 
 
 class TestFractionClass(unittest.TestCase):
@@ -51,17 +51,3 @@ class TestFractionClass(unittest.TestCase):
     def test_can_print_fraction(self):
         frac = Fraction(5, 7)
         self.assertEqual(str(frac), '5/7')
-
-
-class TestMathFunctions(unittest.TestCase):
-    def test_GCD_1_1(self):
-        self.assertEqual(GCD(1, 1), 1)
-
-    def test_GCD_2_4(self):
-        self.assertEqual(GCD(2, 4), 2)
-
-    def test_GCD_81_35(self):
-        self.assertEqual(GCD(81, 35), 1)
-
-    def test_LCM_1_2(self):
-        self.assertEqual(LCM(1, 2), 2)
